chore(network): move bootstrap process to its module
