+++ sn_networking/src/driver.rs
@@ -11,6 +11,7 @@ use crate::metrics::NetworkMetrics;
 #[cfg(feature = "open-metrics")]
 use crate::metrics_service::run_metrics_server;
 use crate::{
+    bootstrap::BOOTSTRAP_INTERVAL,
     circular_vec::CircularVec,
     cmd::SwarmCmd,
     error::{Error, Result},
@@ -95,13 +96,6 @@ const NETWORKING_CHANNEL_SIZE: usize = 10_000;
 
 /// Time before a Kad query times out if no response is received
 const KAD_QUERY_TIMEOUT_S: Duration = Duration::from_secs(25);
-
-/// The interval in which kad.bootstrap is called
-const BOOTSTRAP_INTERVAL: Duration = Duration::from_secs(5);
-
-/// Every BOOTSTRAP_CONNECTED_PEERS_STEP connected peer, we step up the BOOTSTRAP_INTERVAL to slow down bootstrapping
-/// process
-const BOOTSTRAP_CONNECTED_PEERS_STEP: u32 = 50;
 
 // Protocol support shall be downward compatible for patch only version update.
 // i.e. versions of `A.B.X` shall be considered as a same protocol of `A.B`
@@ -526,31 +520,7 @@ impl SwarmDriver {
                 },
                 // runs every bootstrap_interval time
                 _ = bootstrap_interval.tick() => {
-
-                    // kad bootstrap process needs at least one peer in the RT be carried out.
-                    let connected_peers = self.swarm.connected_peers().count() as u32;
-                    if !self.bootstrap_ongoing && connected_peers>=1 {
-                        debug!("Trying to initiate bootstrap. Current bootstrap_interval {:?}", bootstrap_interval.period());
-                        match self.swarm.behaviour_mut().kademlia.bootstrap() {
-                            Ok(query_id) => {
-                                debug!(
-                                    "Initiated kad bootstrap process with query id {query_id:?}"
-                                );
-                                self.bootstrap_ongoing = true;
-                            }
-                            Err(err) => {
-                                error!("Failed to initiate kad bootstrap with error: {err:?}")
-                            }
-                        };
-                    }
-                    // increment bootstrap_interval in steps of INITIAL_BOOTSTRAP_INTERVAL every BOOTSTRAP_CONNECTED_PEERS_STEP
-                    let step = connected_peers / BOOTSTRAP_CONNECTED_PEERS_STEP;
-                    let step = std::cmp::max(1, step);
-                    let new_interval = BOOTSTRAP_INTERVAL * step;
-                    if new_interval > bootstrap_interval.period() {
-                        bootstrap_interval = tokio::time::interval(new_interval);
-                        bootstrap_interval.tick().await; // the first tick completes immediately
-                    }
+                    bootstrap_interval = self.run_bootstrap_continuously(bootstrap_interval).await;
                 }
             }
         }
