+++ sn_networking/src/lib.rs
@@ -9,6 +9,7 @@
 #[macro_use]
 extern crate tracing;
 
+mod bootstrap;
 mod circular_vec;
 mod cmd;
 mod driver;
